instrument plugin calls; move perf API to script-manager
